requirements.txt
atpbar
colorama
tld
python-levenshtein
dnspython3
requests_mock
nose